python/requirements.txt
flask==2.0.2
faiss-cpu==1.7.2
torch==1.10.0
transformers==4.11.3
psycopg2==2.9.2
numpy==1.26.4
